requirements.txt

### requirements.txt

```plaintext
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
```

pip~=24.2
filelock~=3.15.4
pytest~=8.3.2
pluggy~=1.5.0
iniconfig~=2.0.0